scrapper-requirements.txt
beautifulsoup4
lxml
numpy
openpyxl
pandas
psycopg2
selenium
apache-airflow